Rewrite ensuring md5sum.txt
Since we were using a lookup to run md5sum on the
files, this only works when building on localhost.

We are now using the find module for listing files
in /extras. The result in addition to the template
loop file list is then fed to a stat loop, which
will save the md5sum. That then replaces the loop
for lineinfile that ensures md5sum.txt is correct.